// src/app/api/models/user.ts
export class User {
    uid: string;
    email: string;
    displayName: string;
    photoURL: string;
    emailVerified: boolean;
    role?: string;
    firstName?: string;
    lastName?: string;
    byline?: string;
    about?: string;
    admin?: boolean;
    active?: boolean;
    UnitNo?: number;
    phone?: string;
 }